[GEOS-8245] Named layer validation for style groups
